requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
discord.py==2.1.0
emoji==2.0.0
frozenlist==1.3.3
idna==3.4
multidict==6.0.3
peewee==3.15.2
pycparser==2.21
PyNaCl==1.5.0
requests==2.28.1
urllib3==1.26.13
yarl==1.8.2
